requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgi-redis==1.4.2
asgiref==1.1.2
attrs==19.2.0
autobahn==17.7.1
Automat==0.6.0
billiard==3.3.0.23
celery==3.1.25
channels==1.1.6
constantly==15.1.0
daphne==1.3.0
Django==1.8.19
django-bootstrap3==8.1.0
django-celery==3.3.1
hyperlink==17.3.0
incremental==17.5.0
kombu==3.0.37
msgpack-python==0.4.8
packaging==16.8
PyHamcrest==1.10.1
pyparsing==2.2.0
pytz==2016.10
redis==2.10.5
simplejson==3.10.0
six==1.10.0
Twisted==20.3.0
txaio==2.8.1
zope.interface==4.4.2
djangorestframework==3.6.4